+++ src/OpenCOVER/plugins/hlrs/GhostAvatar/controls/BoneParser.cpp
