+++ src/OpenCOVER/plugins/hlrs/GhostAvatar/controls/GhostAvatarControls.cpp
