+++ src/OpenCOVER/plugins/hlrs/GhostAvatar/controls/PlanarAvatarControls.cpp
