+++ src/OpenCOVER/plugins/hlrs/GhostAvatar/controls/TestAvatarControls.cpp
